Moved refactored tests over to main test file
