+++ tests/DB/QueryBuilderTest.php
@@ -3,9 +3,8 @@
 namespace Message\Cog\Test\DB;
 
 use Message\Cog\DB\QueryBuilder;
-use Mockery as m;
 
-class QueryBuilderTest extends \PHPUnit_Framework_TestCase
+class QueryBuilderTestRefactor extends \PHPUnit_Framework_TestCase
 {
 	private $_builder;
 	private $_parser;
@@ -13,18 +12,63 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function setUp()
 	{
-		$this->_connect = m::mock('Message\Cog\DB\Adapter\ConnectionInterface');
-		$this->_parser  = m::mock('Message\Cog\DB\QueryParser');
+		$this->_connect = $this->getMockBuilder('Message\\Cog\\DB\\Adapter\\ConnectionInterface')->disableOriginalConstructor()->getMock();
+		$this->_parser  = $this->getMockBuilder('Message\\Cog\\DB\\QueryParser')->setConstructorArgs(array($this->_connect))->setMethods(array('escape'))->getMock();
 		$this->_builder = new QueryBuilder($this->_connect, $this->_parser);
 	}
-	
+
+	public function testRun()
+	{
+		$result = $this->getMockBuilder('Message\\Cog\\DB\\Adapter\\ResultInterface')
+			->setMethods(array('fetchArray', 'fetchObject', 'seek', 'getAffectedRows', 'getLastInsertId'))
+			->setConstructorArgs(array(null, $this->_connect))
+			->getMockForAbstractClass()
+									;
+
+		$this->_connect->expects($this->once())
+			->method('query')
+			->will($this->returnValue($result));
+
+		$query = $this->_builder
+			->select("*")
+			->from('table')
+			->getQuery()
+			->run();
+
+	}
+
+	public function testAddParams()
+	{
+		$parser  = $this->getMockBuilder('Message\\Cog\\DB\\QueryParser')->setConstructorArgs(array($this->_connect))->setMethods(array('parse'))->getMock();
+		$builder = new QueryBuilder($this->_connect, $parser);
+
+		// Configure the stub.
+		$parser->method('parse')
+					 ->willReturn(true)
+					 ;
+
+		// Test that getQueryString is calling parse
+		$parser->expects($this->once())
+			->method('parse')
+			->with(
+						"SELECT\n*\nFROM table",
+						['param' => 'value']
+					)
+
+			;
+
+		$builder->select("*")
+			->from('table')
+			->addParams(['param' => 'value']);
+
+		// Generate the querystring which will run parse
+		$builder->getQueryString();
+
+	}
+
 	public function testSelectFromSimple()
 	{
 		$expected = "SELECT * FROM table";
-		$this->_parser->shouldReceive('parse')
-			->once()
-			->passthru()
-		;
 
 		$query = $this->_builder
 			->select("*")
@@ -35,12 +79,9 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
 	}
 
-	public function testMultiipleSelect()
+	public function testMultipleSelect()
 	{
 		$expected = "SELECT col_1, col_2, col_3, col_4 FROM table";
-		$this->_parser->shouldReceive('parse')->once()
-			->passthru()
-		;
 
 		$query = $this->_builder
 			->select("col_1")
@@ -57,9 +98,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testSelectArray()
 	{
 		$expected = "SELECT col_1, col_2, col_3, col_4 FROM table";
-		$this->_parser->shouldReceive('parse')->once()
-			->passthru()
-		;
 
 		$query = $this->_builder
 			->select(["col_1", "col_2", "col_3", "col_4"])
@@ -73,9 +111,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testSelectDistinct()
 	{
 		$expected = "SELECT DISTINCT * FROM table";
-		$this->_parser->shouldReceive('parse')->once()
-			->passthru()
-		;
 
 		$query = $this->_builder
 			->select("*", true)
@@ -88,12 +123,7 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function testWhereSimple()
 	{
-		$this->_parser->shouldReceive('parse')->once()
-			->with('col = variable', [])
-			->andReturn('col = variable');
-
 		$expected = "SELECT * FROM table WHERE col = variable";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -107,25 +137,16 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function testWhereWithVariables()
 	{
-		$this->_parser->shouldReceive('parse')->once()
-			->with('col = ?s', ["string"])
-			->andReturn("col = 'string'");
-
-		$this->_parser->shouldReceive('parse')->once()
-			->with('col2 = ?i', [100])
-			->andReturn('col2 = 100');
-
-		$this->_parser->shouldReceive('parse')->once()
-			->with('col3 = ?s', ['ORstring'])
-			->andReturn("col3 = 'ORstring'");
-
 		$expected = "SELECT * FROM table WHERE col = 'string' AND col2 = 100 OR col3 = 'ORstring'";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
+
+		// Configure the stub.
+    $this->_connect->method('escape')
+         ->will($this->returnArgument(0));
 
 		$query = $this->_builder
 			->select("*")
 			->from('table')
-			->where('col = ?s', ["string"])
+			->where('col = ?s', ['string'])
 			->where('col2 = ?i', [100])
 			->where('col3 = ?s', ['ORstring'], false)
 			->getQueryString()
@@ -137,7 +158,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testGroupBy()
 	{
 		$expected = "SELECT * FROM table GROUP BY col";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 		$query = $this->_builder
 			->select("*")
 			->from('table')
@@ -151,7 +171,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testGroupByMulti()
 	{
 		$expected = "SELECT * FROM table GROUP BY col, col2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 			$query = $this->_builder
 			->select("*")
@@ -167,7 +186,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testGroupByArray()
 	{
 		$expected = "SELECT * FROM table GROUP BY col, col2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -175,15 +193,12 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 			->groupBy(['col', 'col2'])
 			->getQueryString()
 		;
-
-
 		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
 	}
 
 	public function testGroupByArrayMulti()
 	{
 		$expected = "SELECT * FROM table GROUP BY col, col2, col3, col4";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -199,7 +214,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testOrderBy()
 	{
 		$expected = "SELECT * FROM table ORDER BY col";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -214,7 +228,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testOrderByMulti()
 	{
 		$expected = "SELECT * FROM table ORDER BY col, col2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -230,7 +243,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testOrderByArray()
 	{
 		$expected = "SELECT * FROM table ORDER BY col, col2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -245,7 +257,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testOrderByArrayMulti()
 	{
 		$expected = "SELECT * FROM table ORDER BY col, col2, col3, col4";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -261,7 +272,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLimit()
 	{
 		$expected = "SELECT * FROM table LIMIT 2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -276,7 +286,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLimitToFrom()
 	{
 		$expected = "SELECT * FROM table LIMIT 2, 5";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -291,7 +300,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testOrderGroupLimit()
 	{
 		$expected = "SELECT * FROM table GROUP BY col, col1 ORDER BY col2, col3 LIMIT 2";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -310,7 +318,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testFromAlias()
 	{
 		$expected = "SELECT * FROM from_table alias";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -330,7 +337,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM (SELECT * FROM from_table) alias";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -362,7 +368,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM (SELECT * FROM table_a UNION SELECT * FROM table_b) alias";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -375,12 +380,7 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function testHaving()
 	{
-		$this->_parser->shouldReceive('parse')->once()
-			->with('col = variable', [])
-			->passthru();
-
 		$expected = "SELECT * FROM table HAVING col = variable";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -394,11 +394,7 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function testHavingMulti()
 	{
-		$this->_parser->shouldReceive('parse')->times(3)
-			->passthru();
-
 		$expected = "SELECT * FROM table HAVING col_a = a AND col_b = b OR col_c = c";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -414,25 +410,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 
 	public function testGetQuery()
 	{
-		$result = m::mock('Message\Cog\DB\Adapter\ResultInterface');
-		$result
-			->shouldReceive('getAffectedRows')->once()
-			->andReturn([])
-			->shouldReceive('getLastInsertId')->once()
-			->andReturn(1)
-		;
-
-		$this->_connect->shouldReceive('query')->once()
-			->andReturn($result)
-		;
-
-		$this->_parser->shouldReceive('parse')->once()
-			->passthru()
-		;
-
-		$expected = "SELECT * FROM table WHERE col = variable";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
-
 		$query = $this->_builder
 			->select("*")
 			->from('table')
@@ -440,16 +417,13 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 			->getQuery()
 		;
 
-		$query->run();
-		$query = $query->getParsedQuery();
+		$this->assertInstanceOf('Message\\Cog\\DB\\Query', $query);
 
-		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
 	}
 
 	public function testJoinSimple()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -464,7 +438,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinAlias()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -479,7 +452,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinSimpleWithOn()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -494,7 +466,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinAliasWithOn()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -509,7 +480,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinSimpleWithUsing()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -524,7 +494,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinAliasWithUsing()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -539,7 +508,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinOnSimple()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -554,7 +522,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinOnAlias()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -569,7 +536,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinUsingSimple()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -584,7 +550,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinUsingAlias()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -605,7 +570,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -626,7 +590,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -647,54 +610,11 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a JOIN (SELECT * FROM table_b) alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
 			->from('table_a')
 			->join('alias', 'id', $table_a, false)
-			->getQueryString()
-		;
-
-		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
-	}
-
-	public function testJoinOnQB()
-	{
-		$table_a = new QueryBuilder($this->_connect, $this->_parser);
-		$table_a
-			->select('*')
-			->from('table_b')
-		;
-
-		$expected = "SELECT * FROM table_a JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
-
-		$query =  $this->_builder
-			->select("*")
-			->from('table_a')
-			->joinOn('alias', 'id = id', $table_a)
-			->getQueryString()
-		;
-
-		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
-	}
-
-	public function testJoinUsingQB()
-	{
-		$table_a = new QueryBuilder($this->_connect, $this->_parser);
-		$table_a
-			->select('*')
-			->from('table_b')
-		;
-
-		$expected = "SELECT * FROM table_a JOIN (SELECT * FROM table_b) alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
-
-		$query =  $this->_builder
-			->select("*")
-			->from('table_a')
-			->joinUsing('alias', 'id', $table_a)
 			->getQueryString()
 		;
 
@@ -728,7 +648,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinSimple()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -737,14 +656,12 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 			->getQueryString()
 		;
 
-
 		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
 	}
 
 	public function testLeftJoinAlias()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -756,11 +673,9 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		$this->assertEquals($expected, trim(preg_replace('/\s+/', ' ', $query)));
 	}
 
-
 	public function testLeftJoinSimpleWithOn()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -775,7 +690,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinAliasWithOn()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -790,7 +704,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinSimpleWithUsing()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -805,7 +718,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinAliasWithUsing()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -820,7 +732,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinOnSimple()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -835,7 +746,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinOnAlias()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -850,7 +760,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinUsingSimple()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -865,7 +774,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinUsingAlias()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -880,7 +788,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinQB()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$table_a = new QueryBuilder($this->_connect, $this->_parser);
 		$table_a
@@ -907,7 +814,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a LEFT JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -928,7 +834,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a LEFT JOIN (SELECT * FROM table_b) alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -949,7 +854,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a LEFT JOIN (SELECT * FROM table_b) alias ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -970,7 +874,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 		;
 
 		$expected = "SELECT * FROM table_a LEFT JOIN (SELECT * FROM table_b) alias USING (id)";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query =  $this->_builder
 			->select("*")
@@ -1009,7 +912,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testJoinBeforeLeftJoin()
 	{
 		$expected = "SELECT * FROM table_a JOIN table_b ON id = id LEFT JOIN table_c ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -1025,7 +927,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testLeftJoinBeforeJoin()
 	{
 		$expected = "SELECT * FROM table_a LEFT JOIN table_b ON id = id JOIN table_c ON id = id";
-		$this->_parser->shouldReceive('parse')->once()->passthru();
 
 		$query = $this->_builder
 			->select("*")
@@ -1039,8 +940,8 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	}
 
 	/**
-     * @expectedException InvalidArgumentException
-     */
+		 * @expectedException InvalidArgumentException
+		 */
 	public function testNoFromError()
 	{
 		$query =  $this->_builder
@@ -1050,8 +951,8 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	}
 
 	/**
-     * @expectedException InvalidArgumentException
-     */
+		 * @expectedException InvalidArgumentException
+		 */
 	public function testNoSelectError()
 	{
 		$query =  $this->_builder
@@ -1063,7 +964,6 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 	public function testEleanorsQuery()
 	{
 		$forQuerys = [];
-		$this->_parser->shouldReceive('parse')->zeroOrMoreTimes()->passthru();
 		$q = new QueryBuilder($this->_connect, $this->_parser);
 		$forQuerys[] = $q
 			->select('item.created_at AS date')
@@ -1160,92 +1060,92 @@ class QueryBuilderTest extends \PHPUnit_Framework_TestCase
 			->from('all', $fromQuery)
 		;
 
-		$expected = 
-'SELECT
-date AS date,
-SUM(all.net) AS net,
-SUM(all.tax) AS tax,
-SUM(all.gross) AS gross,
-all.type AS `type`,
-all.order_id AS `order_id`,
-all.item_id AS item_id,
-all.product,
-all.option,
-country_id AS country
-FROM (SELECT
-item.created_at AS date,
-(IFNULL(item.net, 0)) AS net,
-(IFNULL(item.tax, 0)) AS tax,
-(IFNULL(item.gross, 0)) AS gross,
-CONCAT(order_summary.type," Sale") AS `type`,
-item.item_id AS item_id,
-item.order_id AS order_id,
-item.product_name AS product,
-item.options AS `option`
-FROM order_item AS item
-JOIN order_summary ON item.order_id = order_summary.order_id
-LEFT JOIN return_item ON return_item.exchange_item_id = item.item_id
-WHERE
-order_summary.status_code >= 0
-AND item.product_id NOT IN (9)
-AND return_item.exchange_item_id IS NULL
-AND item.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
-UNION ALL
-SELECT
-order_summary.created_at AS date,
-(IFNULL(net, 0)) AS net,
-(IFNULL(tax, 0)) AS tax,
-(IFNULL(gross, 0)) AS gross,
-"Shipping In" AS `type`,
-"" AS item_id,
-order_shipping.order_id AS order_id,
-"" AS product,
-"" AS `option`
-FROM order_shipping
-JOIN order_summary ON order_shipping.order_id = order_summary.order_id
-WHERE
-order_summary.status_code >= 0
-AND order_summary.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
-UNION ALL
-SELECT
-completed_at AS date,
-(IFNULL(item.net, 0)) AS net,
-(IFNULL(item.tax, 0)) AS tax,
-(IFNULL(item.gross, 0)) AS gross,
-"Exchange item" AS `type`,
-item.item_id AS item_id,
-item.order_id AS order_id,
-item.product_name AS product,
-item.options AS `option`
-FROM order_item AS item
-JOIN order_summary ON order_item.order_id = order_summary.order_id
-JOIN return_item ON return_item.exchange_item_id = item.item_id
-JOIN order_item_status ois ON ois.item_id = item.item_id AND ois.status_code = 0
-WHERE
-order_summary.status_code >= 0
-AND item.product_id NOT IN (9)
-AND item.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
-UNION ALL
-SELECT
-item.completed_at AS date,
--(IFNULL(net, 0)) AS net,
--(IFNULL(tax, 0)) AS tax,
--(IFNULL(gross, 0)) AS gross,
-"Return" AS `type`,
-item.item_id AS item_id,
-item.order_id AS order_id,
-item.product_name AS product,
-item.options AS `option`
-FROM return_item AS item
-JOIN `return` ON return_item.return_id = return.order_id
-WHERE
-accepted = 1
-AND status_code >= 2200
-AND item.product_id NOT IN (9)
-AND item.completed_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())) all';
+		$expected =
+								'SELECT
+								date AS date,
+								SUM(all.net) AS net,
+								SUM(all.tax) AS tax,
+								SUM(all.gross) AS gross,
+								all.type AS `type`,
+								all.order_id AS `order_id`,
+								all.item_id AS item_id,
+								all.product,
+								all.option,
+								country_id AS country
+								FROM (SELECT
+								item.created_at AS date,
+								(IFNULL(item.net, 0)) AS net,
+								(IFNULL(item.tax, 0)) AS tax,
+								(IFNULL(item.gross, 0)) AS gross,
+								CONCAT(order_summary.type," Sale") AS `type`,
+								item.item_id AS item_id,
+								item.order_id AS order_id,
+								item.product_name AS product,
+								item.options AS `option`
+								FROM order_item AS item
+								JOIN order_summary ON item.order_id = order_summary.order_id
+								LEFT JOIN return_item ON return_item.exchange_item_id = item.item_id
+								WHERE
+								order_summary.status_code >= 0
+								AND item.product_id NOT IN (9)
+								AND return_item.exchange_item_id IS NULL
+								AND item.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
+								UNION ALL
+								SELECT
+								order_summary.created_at AS date,
+								(IFNULL(net, 0)) AS net,
+								(IFNULL(tax, 0)) AS tax,
+								(IFNULL(gross, 0)) AS gross,
+								"Shipping In" AS `type`,
+								"" AS item_id,
+								order_shipping.order_id AS order_id,
+								"" AS product,
+								"" AS `option`
+								FROM order_shipping
+								JOIN order_summary ON order_shipping.order_id = order_summary.order_id
+								WHERE
+								order_summary.status_code >= 0
+								AND order_summary.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
+								UNION ALL
+								SELECT
+								completed_at AS date,
+								(IFNULL(item.net, 0)) AS net,
+								(IFNULL(item.tax, 0)) AS tax,
+								(IFNULL(item.gross, 0)) AS gross,
+								"Exchange item" AS `type`,
+								item.item_id AS item_id,
+								item.order_id AS order_id,
+								item.product_name AS product,
+								item.options AS `option`
+								FROM order_item AS item
+								JOIN order_summary ON order_item.order_id = order_summary.order_id
+								JOIN return_item ON return_item.exchange_item_id = item.item_id
+								JOIN order_item_status ois ON ois.item_id = item.item_id AND ois.status_code = 0
+								WHERE
+								order_summary.status_code >= 0
+								AND item.product_id NOT IN (9)
+								AND item.created_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())
+								UNION ALL
+								SELECT
+								item.completed_at AS date,
+								-(IFNULL(net, 0)) AS net,
+								-(IFNULL(tax, 0)) AS tax,
+								-(IFNULL(gross, 0)) AS gross,
+								"Return" AS `type`,
+								item.item_id AS item_id,
+								item.order_id AS order_id,
+								item.product_name AS product,
+								item.options AS `option`
+								FROM return_item AS item
+								JOIN `return` ON return_item.return_id = return.order_id
+								WHERE
+								accepted = 1
+								AND status_code >= 2200
+								AND item.product_id NOT IN (9)
+								AND item.completed_at BETWEEN UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL 12 MONTH)) AND UNIX_TIMESTAMP(NOW())) all';
 
-		$this->_parser->shouldReceive('parse')->once()->andReturn($expected);
 
 		$this->assertEquals(trim(preg_replace('/\s+/', ' ', $expected)), trim(preg_replace('/\s+/', ' ', $query->getQueryString())));
 	}
+
 }
